requirements.txt
aioodbc
aiohttp
aiohttp_cors
pyodbc
pyjwt
marshmallow



















